massive refactor of shaders, shader loading, and how the immediate api uses assets.
Shaders are now loaded by the engine and provided to the renderer as a render context.

Immediate Mode rendering requires a generic 2d shader to draw stuff.  That must be provided by the initializer of the render library (the engine in this case).

Added support for file loading shaders, and far better shader compiler error reporting.

Added support for zip file packs.

Added support for #inject preprocessor and injecting engine data into shader text.

Implemented the error shader and the generic 2d shader as provided by the engine pack.